# docs/swagger.yaml
basePath: /v1/users
definitions:
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.ClearMessage:
    properties:
      chat_id:
        type: string
      community_type:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.LinkPreviewResponse:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.Message:
    properties:
      attachments:
        type: string
      chat_id:
        type: string
      community_type:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_edited:
        type: boolean
      message:
        type: string
      message_type:
        type: string
      reactions:
        items:
          type: string
        type: array
      reply_to_id:
        type: string
      sender_id:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_gorilla_websocket.Conn:
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is PingVi selective forwarding unit.
  termsOfService: http://swagger.io/terms/
  title: PingVi selective forwarding unit
  version: "1.0"
paths:
...
  /ws/{session_id}:
    get:
      consumes:
      - application/json
      description: |-
        Establishes a WebSocket connection for real-time messaging.
        - Requires session_id for authentication
        - Supports operations: switch_community, read, send, update, delete
        - Messages are broadcasted to all members of the chat
      parameters:
      - description: Session ID for authentication
        example: '"abc123def456"'
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: WebSocket connection upgraded
          schema:
            $ref: '#/definitions/github_com_gorilla_websocket.Conn'
        "400":
          description: "Invalid session ID\"\t\texample({\"error\":\"invalid session
            id\"})"
          schema:
            additionalProperties: true
            type: object
        "415":
          description: "Invalid content type\"\t\texample({\"error\":\"Content-Type
            must be application/json\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "WebSocket upgrade failed\"\texample({\"error\":\"WebSocket
            upgrade failed\"})"
          schema:
            additionalProperties: true
            type: object
      summary: WebSocket connection for real-time messaging
      tags:
      - messenger-websocket
  /ws/messages/{chat_id}:
    get:
      consumes:
      - application/json
      description: Retrieves all messages from a specific chat (channel, group, or
        personal chat).
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Chat ID
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: chat_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of messages
          schema:
            items:
              $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.Message'
            type: array
        "400":
          description: "Invalid chat ID\"\t\texample({\"error\":\"invalid chat id\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\texample({\"error\":\"failed to get
            messages\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all messages from a chat
      tags:
      - messenger-websocket
  /ws/messages/clear:
    delete:
      consumes:
      - application/json
      description: Clears all messages from a specific chat. Requires appropriate
        permissions.
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Chat ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.ClearMessage'
      produces:
      - application/json
      responses:
        "200":
          description: "Chat ID\"\t\t\t\texample({\"chat_id\":\"123e4567-e89b-12d3-a456-426614174000\"})"
          schema:
            additionalProperties: true
            type: object
        "400":
          description: "Invalid request\"\t\texample({\"error\":\"invalid body\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\texample({\"error\":\"failed to clear
            messages\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Clear all messages from a chat
      tags:
      - messenger-websocket
  /ws/preview/{link}:
    get:
      consumes:
      - application/json
      description: |-
        Fetches a preview for a given link. Supports internal PingVi links and external URLs.
        - Internal links: returns type and message
        - External links: fetches content type and returns preview
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: URL to preview
        example: '"https://example.com/image.jpg"'
        in: path
        name: link
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: "Link preview data\"\t\texample({\"type\":\"image\",\"message\":\"https://s3.amazonaws.com/public/emoji/standart/1F600.json\"})"
          schema:
            $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.LinkPreviewResponse'
        "400":
          description: "Invalid URL\"\t\t\texample({\"error\":\"invalid url format\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\texample({\"error\":\"failed to fetch
            preview\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get link preview
      tags:
      - messenger-websocket
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"